+++ src/Laravel/ApiPlatformProvider.php
@@ -744,6 +744,11 @@ class ApiPlatformProvider extends ServiceProvider
                 oAuthRefreshUrl: $config->get('api-platform.swagger_ui.oauth.refreshUrl', null),
                 oAuthScopes: $config->get('api-platform.swagger_ui.oauth.scopes', []),
                 apiKeys: $config->get('api-platform.swagger_ui.apiKeys', []),
+                contactName: $config->get('api-platform.swagger_ui.contact.name', ''),
+                contactUrl: $config->get('api-platform.swagger_ui.contact.url', ''),
+                contactEmail: $config->get('api-platform.swagger_ui.contact.email', ''),
+                licenseName: $config->get('api-platform.swagger_ui.license.name', ''),
+                licenseUrl: $config->get('api-platform.swagger_ui.license.url', ''),
             );
         });
 
@@ -1130,14 +1135,18 @@ class ApiPlatformProvider extends ServiceProvider
 
         $app->singleton('api_platform.graphql.type_locator', function (Application $app) {
             $tagged = iterator_to_array($app->tagged('api_platform.graphql.type'));
+            $services = [];
+            foreach ($tagged as $service) {
+                $services[$service->name] = $service;
+            }
 
-            return new ServiceLocator($tagged);
+            return new ServiceLocator($services);
         });
 
         $app->singleton(TypesFactoryInterface::class, function (Application $app) {
             $tagged = iterator_to_array($app->tagged('api_platform.graphql.type'));
 
-            return new TypesFactory($app->make('api_platform.graphql.type_locator'), array_keys($tagged));
+            return new TypesFactory($app->make('api_platform.graphql.type_locator'), array_column($tagged, 'name'));
         });
         $app->singleton(TypesContainerInterface::class, function () {
             return new TypesContainer();